separate looper from watcher
